Use long for Channel#await duration
This matches the underlying method called by AbstractChannel and is the
standard for timeouts with a TimeUnit.

+++ src/main/java/net/schmizz/sshj/connection/channel/Channel.java
@@ -131,7 +131,7 @@ public interface Channel
     void join()
             throws ConnectionException;
 
-    void join(int timeout, TimeUnit unit)
+    void join(long timeout, TimeUnit unit)
             throws ConnectionException;
 
 }
